requirements.txt
flask==2.3.2
python-dotenv
langchain-google-genai
langchain-core